requirements.txt
typing-inspect==0.8.0
typing_extensions==4.5.0
openapi-schema-pydantic==1.2.4
pydantic==1.10.11
python-dotenv==1.0.0
weaviate-client==3.25.3
openai==1.3.7
langchain==0.0.344
tiktoken==0.5.1
streamlit==1.29.0
streamlit-chat==0.1.1
pypdf==3.17.1
